solve the problem #844

+++ src/ACMP844.java
@@ -4,9 +4,9 @@ import java.util.Scanner;
 public class ACMP844 {
     // your solution here
     private void solve(Scanner in, PrintWriter out) {
-        int a = in.nextInt();
-        int b = in.nextInt();
-        int s = a * b;
+        long a = in.nextLong();
+        long b = in.nextLong();
+        long s = a * b;
         double side = Math.sqrt(s);
 
         if (side % 1 == 0) {
